Sort streams by type in raw view
The method minidump::Minidump::all_streams iterates over a HashMap which
is documented to return values in an arbitrary order.

In the UI, that can be a little confusing, so I added sorting based on
the stream type which is an integer value. I chose this order over the
names because the more commonly used streams seem to appear first when
sorted by type.